feat: extract segments selection table
