+++ src/frontend/src/lib/components/core/SegmentsTable.svelte
@@ -1,0 +1,172 @@
+<script lang="ts">
+	import type { Principal } from '@dfinity/principal';
+	import { nonNullish } from '@dfinity/utils';
+	import { onMount, type Snippet } from 'svelte';
+	import type { Orbiter, Satellite } from '$declarations/mission_control/mission_control.did';
+	import { satellitesStore } from '$lib/derived/satellite.derived';
+	import { loadOrbiters } from '$lib/services/orbiters.services';
+	import { loadSatellites } from '$lib/services/satellites.services';
+	import { i18n } from '$lib/stores/i18n.store';
+	import { orbiterStore } from '$lib/stores/orbiter.store';
+	import { orbiterName } from '$lib/utils/orbiter.utils';
+	import { satelliteName } from '$lib/utils/satellite.utils';
+
+	interface Props {
+		missionControlId: Principal;
+		children?: Snippet;
+		selectedMissionControl: boolean;
+		selectedSatellites: [Principal, Satellite][];
+		selectedOrbiters: [Principal, Orbiter][];
+	}
+
+	let {
+		missionControlId,
+		children,
+		selectedMissionControl = $bindable(false),
+		selectedSatellites = $bindable([]),
+		selectedOrbiters = $bindable([])
+	}: Props = $props();
+
+	let satellites: [Principal, Satellite][] = $state([]);
+	let orbiters: [Principal, Orbiter][] = $state([]);
+
+	const loadSegments = async () => {
+		const [{ result: resultSatellites }, { result: resultOrbiters }] = await Promise.all([
+			loadSatellites({ missionControl: missionControlId, reload: true }),
+			loadOrbiters({ missionControl: missionControlId, reload: true })
+		]);
+
+		if (resultSatellites !== 'success' || resultOrbiters !== 'success') {
+			satellites = [];
+			orbiters = [];
+			return;
+		}
+
+		satellites = ($satellitesStore ?? []).map(({ satellite_id, ...rest }) => [
+			satellite_id,
+			{ satellite_id, ...rest }
+		]);
+		orbiters = nonNullish($orbiterStore) ? [[$orbiterStore.orbiter_id, $orbiterStore]] : [];
+
+		toggleAll();
+	};
+
+	onMount(() => {
+		loadSegments();
+	});
+
+	let allSelected = $state(false);
+
+	const toggleAll = () => {
+		allSelected = !allSelected;
+
+		selectedMissionControl = allSelected;
+		selectedSatellites = allSelected ? [...satellites] : [];
+		selectedOrbiters = allSelected ? [...orbiters] : [];
+	};
+</script>
+
+<div class="table-container">
+	{@render children?.()}
+
+	<table>
+		<thead>
+			<tr>
+				<th class="tools"> {$i18n.cli.selected} </th>
+				<th> {$i18n.cli.module} </th>
+			</tr>
+		</thead>
+		<tbody>
+			<tr>
+				<td class="actions"><input type="checkbox" bind:checked={selectedMissionControl} /></td>
+
+				<td>
+					<span>{$i18n.mission_control.title}</span>
+					<span class="canister-id">({missionControlId.toText()})</span>
+				</td>
+			</tr>
+
+			{#each satellites as satellite}
+				<tr>
+					<td class="actions"
+						><input type="checkbox" bind:group={selectedSatellites} value={satellite} /></td
+					>
+					<td
+						><span>{satelliteName(satellite[1])}</span>
+						<span class="canister-id">({satellite[0].toText()})</span></td
+					>
+				</tr>
+			{/each}
+
+			{#each orbiters as orbiter}
+				{@const orbName = orbiterName(orbiter[1])}
+
+				<tr>
+					<td class="actions"
+						><input type="checkbox" bind:group={selectedOrbiters} value={orbiter} /></td
+					>
+					<td>
+						<span>{!orbName ? $i18n.analytics.title : orbName}</span>
+						<span class="canister-id">({orbiter[0].toText()})</span>
+					</td>
+				</tr>
+			{/each}
+		</tbody>
+	</table>
+</div>
+
+<div class="objects">
+	<div class="checkbox all">
+		<input type="checkbox" onchange={toggleAll} checked={allSelected} />
+		<span>{allSelected ? $i18n.core.unselect_all : $i18n.core.select_all}</span>
+	</div>
+</div>
+
+<style lang="scss">
+	@use '../../../lib/styles/mixins/text';
+
+	.tools {
+		width: 88px;
+	}
+
+	.checkbox {
+		display: flex;
+		align-items: center;
+		gap: var(--padding-2x);
+	}
+
+	span {
+		@include text.truncate;
+	}
+
+	.canister-id {
+		font-size: var(--font-size-ultra-small);
+	}
+
+	.all {
+		display: flex;
+		align-items: flex-start;
+
+		font-size: var(--font-size-ultra-small);
+
+		margin: var(--padding) 0 0;
+	}
+
+	.objects {
+		margin: 0 0 var(--padding-4x);
+		padding: var(--padding) var(--padding-2x);
+	}
+
+	.table-container {
+		margin: 0;
+	}
+
+	.actions {
+		display: flex;
+		padding: var(--padding-2x) var(--padding-2x);
+	}
+
+	input[type='checkbox'] {
+		margin: 0;
+	}
+</style>
