+++ src/frontend/src/lib/i18n/zh-cn.json
@@ -62,7 +62,9 @@
 		"hold_tight": "稍等!",
 		"create": "创建",
 		"replace": "替换",
-		"restore": "恢复"
+		"restore": "恢复",
+		"select_all": "选择全部",
+		"unselect_all": "取消全部选择"
 	},
 	"canisters": {
 		"insight": "分析",
@@ -413,8 +415,6 @@
 		"selected": "已选择",
 		"module": "模块",
 		"terminal": "终端",
-		"select_all": "选择全部",
-		"unselect_all": "取消全部选择",
 		"profile_info": "如果需要，提供一个简介名字有助于在未来更容易识别.",
 		"segments": "模块:",
 		"profile_placeholder": "可选的简介名字",
